create api for document category

+++ migrations/20200211061025-create-document-list.js
@@ -9,10 +9,10 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       category_id: {
-        type: Sequelize.INTEGER(11)
+        type: Sequelize.INTEGER
       },
       sub_category_id: {
-        type: Sequelize.INTEGER(11)
+        type: Sequelize.INTEGER
       },
       title: {
         type: Sequelize.STRING(127)
